+++ lib/src/layers/fill_extrusion_layer.dart
@@ -217,10 +217,11 @@ class FillExtrusionLayerProperties extends LayerProperties {
   /// Accepted data type - String
   final dynamic sourceLayer;
 
-  /// Whether this layer is displayed.
-  /// Accepted data type - bool
-  /// default value is true
-  final dynamic visibility;
+  /// Whether this layer should be visible or not.
+  /// Accepted data type - [LayerVisibility]
+  /// default value is LayerVisibility.visible
+  ///
+  final LayerVisibility visibility;
 
   /// The minimum zoom level for the layer. At zoom levels less than
   /// the min-zoom, the layer will be hidden.
@@ -262,7 +263,7 @@ class FillExtrusionLayerProperties extends LayerProperties {
     this.fillExtrusionVerticalGradient,
     this.filter,
     this.sourceLayer,
-    this.visibility,
+    this.visibility = LayerVisibility.visible,
     this.minZoom,
     this.maxZoom,
   });
@@ -321,7 +322,7 @@ class FillExtrusionLayerProperties extends LayerProperties {
       }
     }
 
-    insert('visibility', visibility);
+    insert('visibility', visibility.name);
 
     return layoutArgs;
   }