+++ lib/src/layers/fill_layer.dart
@@ -159,10 +159,11 @@ class FillLayerProperties extends LayerProperties {
   /// Accepted data type - Expression
   final dynamic filter;
 
-  /// Whether this layer is displayed.
-  /// Accepted data type - bool
-  /// default value is true
-  final dynamic visibility;
+  /// Whether this layer should be visible or not.
+  /// Accepted data type - [LayerVisibility]
+  /// default value is LayerVisibility.visible
+  ///
+  final LayerVisibility visibility;
 
   /// The minimum zoom level for the layer. At zoom levels less than
   /// the min-zoom, the layer will be hidden.
@@ -200,7 +201,7 @@ class FillLayerProperties extends LayerProperties {
     this.filter,
     this.maxZoom,
     this.minZoom,
-    this.visibility,
+    this.visibility = LayerVisibility.visible,
   });
 
   /// Default fill layer properties
@@ -258,7 +259,7 @@ class FillLayerProperties extends LayerProperties {
     }
 
     insert('fill-sort-key', fillSortKey);
-    insert('visibility', visibility);
+    insert('visibility', visibility.name);
 
     return layoutArgs;
   }