+++ lib/src/layers/heatmap_layer.dart
@@ -128,10 +128,11 @@ class HeatmapLayerProperties extends LayerProperties {
   /// Accepted data type - String
   final dynamic sourceLayer;
 
-  /// Whether this layer is displayed.
-  /// Accepted data type - bool
-  /// default value is true
-  final dynamic visibility;
+  /// Whether this layer should be visible or not.
+  /// Accepted data type - [LayerVisibility]
+  /// default value is LayerVisibility.visible
+  ///
+  final LayerVisibility visibility;
 
   /// The minimum zoom level for the layer. At zoom levels less than
   /// the min-zoom, the layer will be hidden.
@@ -163,7 +164,7 @@ class HeatmapLayerProperties extends LayerProperties {
     this.heatmapWeight,
     this.filter,
     this.sourceLayer,
-    this.visibility,
+    this.visibility = LayerVisibility.visible,
     this.minZoom,
     this.maxZoom,
   });
@@ -223,7 +224,7 @@ class HeatmapLayerProperties extends LayerProperties {
       }
     }
 
-    insert('visibility', visibility);
+    insert('visibility', visibility.name);
 
     return layoutArgs;
   }