+++ lib/src/layers/hill_shade_layer.dart
@@ -108,10 +108,11 @@ class HillShadeLayerProperties extends LayerProperties {
   /// Accepted data type - String
   final dynamic sourceLayer;
 
-  /// Whether this layer is displayed.
-  /// Accepted data type - bool
-  /// default value is true
-  final dynamic visibility;
+  /// Whether this layer should be visible or not.
+  /// Accepted data type - [LayerVisibility]
+  /// default value is LayerVisibility.visible
+  ///
+  final LayerVisibility visibility;
 
   /// The minimum zoom level for the layer. At zoom levels less than
   /// the min-zoom, the layer will be hidden.
@@ -144,7 +145,7 @@ class HillShadeLayerProperties extends LayerProperties {
     this.hillShadeShadowColor,
     this.hillShadeShadowColorTransition,
     this.sourceLayer,
-    this.visibility,
+    this.visibility = LayerVisibility.visible,
     this.minZoom,
     this.maxZoom,
   });
@@ -203,7 +204,7 @@ class HillShadeLayerProperties extends LayerProperties {
       }
     }
 
-    insert('visibility', visibility);
+    insert('visibility', visibility.name);
 
     return layoutArgs;
   }