+++ lib/src/layers/raster_layer.dart
@@ -126,10 +126,11 @@ class RasterLayerProperties extends LayerProperties {
   /// Accepted data type - String
   final dynamic sourceLayer;
 
-  /// Whether this layer is displayed.
-  /// Accepted data type - bool
-  /// default value is true
-  final dynamic visibility;
+  /// Whether this layer should be visible or not.
+  /// Accepted data type - [LayerVisibility]
+  /// default value is LayerVisibility.visible
+  ///
+  final LayerVisibility visibility;
 
   /// The minimum zoom level for the layer. At zoom levels less than
   /// the min-zoom, the layer will be hidden.
@@ -166,7 +167,7 @@ class RasterLayerProperties extends LayerProperties {
     this.rasterSaturation,
     this.rasterSaturationTransition,
     this.sourceLayer,
-    this.visibility,
+    this.visibility = LayerVisibility.visible,
     this.minZoom,
     this.maxZoom,
   });
@@ -225,7 +226,7 @@ class RasterLayerProperties extends LayerProperties {
       }
     }
 
-    insert('visibility', visibility);
+    insert('visibility', visibility.name);
 
     return layoutArgs;
   }